+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 - check-ssl-host.rb: Added optional `address` command line parameter for specifying the address of the server to
                      connect to, to override the `hostname` parameter (which is still used for verification/SNI)
 
+- check-ssl-host.rb: Basic IMAP STARTTLS negotiation
+
 ## [1.0.0]
 ### Changed
 - Updated Rubocop to 0.40, applied auto-correct